Fe board (#47)
* modify : security 추가검토

* modify : etc

* modify : 문의사항 완료
add : 공지사항(파일업로드 구현중)

---------

+++ src/main/java/com/myrium/domain/AuthVO.java
@@ -5,8 +5,7 @@ import lombok.Data;
 @Data
 public class AuthVO {
 
-	private Integer userId;
+	private Long userId;
 	private String role;
-	
 	private String customerId;
 }
